Admin role

+++ Data_Layer/Interfaces/IRatingService.cs
@@ -7,7 +7,7 @@ namespace Interfaces
         public int AddRating(RatingDTO Rating);
         public List<RatingDTO> GetRatingsByGame(int gameId);
         public bool UserHasRated(int userId, int gameId);
-        public int GetAverageScoreFromGames(int gameId);
+        public double GetAverageScoreFromGames(int gameId);
 
     }
 }
